smaller NSP installer
