--- source/NSP/extra/unzip.hpp
@@ -1,3 +1,0 @@
-namespace inst::zip {
-    bool extractFile(const std::string filename, const std::string destination);
-}
